refactor: add usePressingAnimation hook
